abstract function for event create
